+++ src/TaskOption.ts
@@ -7,7 +7,10 @@
 import { TaskOption } from "fp-ts/TaskOption"
 import * as T from "fp-ts/Task"
 import { execute as executeT } from "./Task"
-import { unsafeUnwrap as unsafeUnwrapO } from "./Option"
+import {
+  unsafeUnwrap as unsafeUnwrapO,
+  unsafeExpect as unsafeExpectO,
+} from "./Option"
 import { flow } from "fp-ts/function"
 
 /**
@@ -27,3 +30,23 @@ export const unsafeUnwrap: <A>(x: TaskOption<A>) => Promise<A> = flow(
   T.map(unsafeUnwrapO),
   executeT,
 )
+
+/**
+ * Unwrap the promise from within a `TaskOption`, rejecting with `msg` if
+ * `None`.
+ *
+ * @example
+ * import { unsafeExpect } from 'fp-ts-std/TaskOption'
+ * import * as TO from 'fp-ts/TaskOption'
+ *
+ * assert.rejects(
+ *   unsafeExpect('foo')(TO.none),
+ *   /^foo$/,
+ * )
+ *
+ * @since 0.16.0
+ */
+export const unsafeExpect = (
+  msg: string,
+): (<A>(x: TaskOption<A>) => Promise<A>) =>
+  flow(T.map(unsafeExpectO(msg)), executeT)