+++ test/TaskOption.ts
@@ -1,4 +1,4 @@
-import { unsafeUnwrap } from "../src/TaskOption"
+import { unsafeUnwrap, unsafeExpect } from "../src/TaskOption"
 import * as TO from "fp-ts/TaskOption"
 
 describe("TaskOption", () => {
@@ -11,6 +11,18 @@ describe("TaskOption", () => {
 
     it("throws None", () => {
       expect(() => f(TO.none)).rejects
+    })
+  })
+
+  describe("unsafeExpect", () => {
+    const f = unsafeExpect("foo")
+
+    it("unwraps Some", () => {
+      return expect(f(TO.some(123))).resolves.toBe(123)
+    })
+
+    it("throws None with provided message", () => {
+      return expect(() => f(TO.none)).rejects.toBe("foo")
     })
   })
 })